+++ ONE-PIX_soft/src/reconstructions_methods/FourierShiftReconstruction.py
@@ -42,7 +42,7 @@ class FourierShiftReconstruction:
         y=np.asarray(y)     
         spectre_desplit=np.asarray(spectre_desplit)
         # spectre_desplit=self.snr_filt(spectre_desplit)
-        half_spectrum=np.zeros((np.max(x)+1,np.max(y)+1,np.size(spectre_desplit,1)),dtype=np.complex128)
+        half_spectrum=np.zeros((np.max(x)+1,np.max(y)+1,np.size(spectre_desplit,1)),dtype=np.complex64)
         
         for i in range(0,np.size(x)):
             half_spectrum[x[i],y[i]]=spectre_desplit[i,:]
