+++ ONE-PIX_soft/src/reconstructions_methods/HadamardReconstruction.py
@@ -38,7 +38,7 @@ class HadamardReconstruction:
         y=np.asarray(y)     
         spectre_desplit=np.asarray(spectre_desplit)
         
-        whole_spectrum=np.zeros((np.max(x)+1,np.max(y)+1,np.size(spectre_desplit,1)))
+        whole_spectrum=np.zeros((np.max(x)+1,np.max(y)+1,np.size(spectre_desplit,1)),dtype=np.float32)
         
         for i in range(0,np.size(x)):
             whole_spectrum[x[i],y[i]]=spectre_desplit[i,:]
